requirements.txt
streamlit
faiss-cpu
numpy
pypdf
sentence-transformers
python-dotenv
openai